ST-5 User profile \ Correct small things.
Methods User::createNew and Profile::updateUserData moved from models into services to improve code readability. Used radiobattons instead text input to limit the input options.
