+++ resources/views/profiles/edit.blade.php
@@ -2,7 +2,7 @@
 @section('title', 'Редактировать профиль')
 
 @section('content')
-    <body>
+
     <h1>Редактировать профиль</h1>
     <hr>
     <form method="post" action="{{ route('profiles.show', ['profile' => $profile]) }}">
